+++ test/unitcore/test_basedal.py
@@ -78,6 +78,8 @@ class test_datastore_voodooentry_point(unittest.TestCase):
             self.subject.loads('payload', format=1)
         with self.assertRaises(NotImplementedError):
             self.subject.dumps(format=1)
+        with self.assertRaises(NotImplementedError):
+            self.subject.get_raw_xpath('/xpath', with_val=False)
         self.subject.setup_root()
         self.subject._initdal()
         """