+++ yangvoodoo/basedal.py
@@ -221,3 +221,9 @@ class BaseDataAbstractionLayer:
 
     def merges(self, filename, format=1, trusted=True):
         raise NotImplementedError("merges not implemented")
+
+    def get_raw_xpath(self, xpath, with_val=False):
+        raise NotImplementedError("get_raw_xpath not implemented")
+
+    def get_raw_xpath_single_val(self, xpath):
+        raise NotImplementedError("get_raw_xpath_single_val not implemented")