+++ CHANGELOG.md
@@ -101,3 +101,5 @@
    - If a DAL returns None for a number looking type return 0 instead.
    - Deprecated TemplateNinja
    - Fix for matching uint/decimal64 in a union (the dal returns a string not a python int)
+   - Add `get_raw_xpath` to sublydal
+   - Add get_index(n) to a List to return the item at position if it exists.